Added the get-kernel option, which lets a user download a suitable kernel.
Extended the instruction in the readme file, documenting the get-kernel option, and adding a tutorial/example from start to finish.

in pkglib:sub set - bugfix - if you change a setting on an existing package, note down the package that contain the file.

Packages now know the path to the deb files, so they can examine them.